+++ enclone_visual/src/gui.rs
@@ -84,8 +84,6 @@ impl Application for EncloneVisual {
     // ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
 
     fn view(&mut self) -> Element<Message> {
-        const SCROLLBAR_WIDTH: u16 = 12;
-        const SPACING: u16 = 20;
         let version = VERSION.lock().unwrap()[0].clone();
         let version_float = format!("1e-{}", -version.force_f64().log10());
 
@@ -94,33 +92,7 @@ impl Application for EncloneVisual {
         // Handle the summary case.
 
         if self.summary_mode {
-            let summary_title = Text::new(&format!("Summary")).size(30);
-            let summary = SUMMARY_CONTENTS.lock().unwrap()[0].clone();
-            let nlines = summary.chars().filter(|&n| n == '\n').count();
-            let font_size = (15 * nlines) / 38;
-            let summary_scrollable = Scrollable::new(&mut self.scroll)
-                .width(Length::Fill)
-                .height(Length::Fill)
-                .scrollbar_width(SCROLLBAR_WIDTH)
-                .scroller_width(12)
-                .style(style::ScrollableStyle)
-                .push(
-                    Text::new(&format!("{}", summary))
-                        .font(DEJAVU_BOLD)
-                        .size(font_size as u16),
-                );
-            let summary_close_button = Button::new(&mut self.open_state, Text::new("Vanish!"))
-                .on_press(Message::SummaryClose);
-            let content = Column::new()
-                .spacing(SPACING)
-                .padding(20)
-                .push(summary_title)
-                .push(summary_scrollable)
-                .push(summary_close_button);
-            return Container::new(content)
-                .width(Length::Fill)
-                .height(Length::Fill)
-                .into();
+            return summary(self);
         }
 
         // ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
